\name{current_treaties_dyads}
\alias{current_treaties_dyads}
\docType{data}
\title{
  Current PTAs in Dyadic Form
}
\description{
This table does not come with the DESTA downloads. I've computed a binary
 variable that expresses if on year "Y", two countries, expressed as dyads
 in alphabetical order, have at least one valid PTA in common. This emerges
 from the fact that some countries may retire from previous trade agreements.
}
\usage{current_treaties_dyads}
\format{
  A data frame with 473,674 observations.
}
\source{
Based on the datasets from https://www.designoftradeagreements.org/.
}
\references{
Dür, Andreas, Leonardo Baccini and Manfred Elsig. 2014. "The Design of International Trade Agreements: Introducing a New Database". The Review of International Organizations, 9(3): 353-375.
}
\examples{
current_treaties_dyads
}
\keyword{datasets}
