\name{in_force_ftas_dyads}
\alias{in_force_ftas_dyads}
\docType{data}
\title{
  Current FTAs in Dyadic Form
}
\description{
This table does not come with the DESTA downloads. I've computed a binary
 variable that expresses if on year "Y", two countries, expressed as dyads
 in alphabetical order, have at least one in force common Full FTA. This emerges
 from the fact that some countries may retire from previous trade agreements.
}
\usage{in_force_ftas_dyads}
\format{
  A data frame with 269,500 observations.
}
\source{
Based on the datasets from https://www.designoftradeagreements.org/.
}
\references{
Dür, Andreas, Leonardo Baccini and Manfred Elsig. 2014. "The Design of International Trade Agreements: Introducing a New Database". The Review of International Organizations, 9(3): 353-375.
}
\examples{
in_force_ftas_dyads
}
\keyword{datasets}
