\name{treaties}
\alias{treaties}
\docType{data}
\title{
  List of Treaties
}
\description{
Partial scope agreements are included as soon as they liberalize at least some
 trade, whereas framework agreements (with very few exceptions), trade and
 co-operation agreements, etc. are excluded. Does not include interim
 agreements (for this reason, no agreements with the Palestinian Authority
 are included). Neither considers agreements with tiny island states such as
 the Marshall Islands or Tuvalu.
}
\usage{treaties}
\format{
  A data frame with 1,340 observations indicating agreements and their
  characteristics. Please see the detailed description
  in the DESTA Codebook and Indices. Both files can be obtained from
  https://www.designoftradeagreements.org/downloads/.
}
\source{
Downloaded from https://www.designoftradeagreements.org/.
}
\references{
Dür, Andreas, Leonardo Baccini and Manfred Elsig. 2014. "The Design of International Trade Agreements: Introducing a New Database". The Review of International Organizations, 9(3): 353-375.
}
\examples{
treaties
}
\keyword{datasets}
