\name{withdrawals_dyads}
\alias{withdrawals_dyads}
\docType{data}
\title{
  PTAs Withdrawals in Dyadic Form
}
\description{
Withdrawals from previously subscribed agreements, which were are included as
 soon as they liberalize at least some trade, whereas framework agreements
 (with very few exceptions), trade and co-operation agreements, etc. are
 excluded. Does not include interim agreements (for this reason, no agreements
 with the Palestinian Authority are included). Neither considers agreements
 with tiny island states such as the Marshall Islands or Tuvalu.
}
\usage{withdrawals_dyads}
\format{
  A data frame with 1,690 observations indicating all the pairs involved in
  the different withdrawals from trade agreements. Please see the detailed
  description in the DESTA Codebook and Indices. Both files can be obtained from
  https://www.designoftradeagreements.org/downloads/.
}
\source{
Downloaded from https://www.designoftradeagreements.org/.
}
\references{
Dür, Andreas, Leonardo Baccini and Manfred Elsig. 2014. "The Design of International Trade Agreements: Introducing a New Database". The Review of International Organizations, 9(3): 353-375.
}
\examples{
withdrawals_dyads
}
\keyword{datasets}
